# backend/src/main/resources/openapi/openapi.yaml
openapi: "3.0.3"
info:
  title: "timetable API"
  description: "timetable API"
  version: "1.0.0"
servers:
  - url: "https://t"
paths:
...
  /api/v1/admin/audience/by-id:
    get:
      summary: "GET api/v1/admin/audience/by-id"
      operationId: "getAudienceById"
      parameters:
        - name: "id"
          in: "query"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AudienceDTO"
      tags:
        - audience
  /api/v1/admin/audience/by-name:
    get:
      summary: "GET api/v1/admin/audience/by-name"
      operationId: "getAudienceByName"
      parameters:
        - name: "name"
          in: "query"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AudienceDTO"
      tags:
        - audience
...
  /api/v1/admin/audience:
    post:
      summary: "POST api/v1/admin/audience"
      operationId: "addAudience"
      parameters:
        - name: "name"
          in: "query"
          required: true
          schema:
            type: "string"
        - name: "audience_type"
          in: "query"
          required: true
          schema:
            type: "string"
            enum:
              - "LECTURE"
              - "TERMINAL"
              - "PRACTICAL"
              - "LABORATORY"
        - name: "audience_group_capacity"
          in: "query"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AudienceDTO"
      tags:
        - audience
    put:
      summary: "PUT api/v1/admin/audience"
      operationId: "updateAudience"
      parameters:
        - name: "current"
          in: "query"
          required: true
          schema:
            type: "string"
        - name: "new"
          in: "query"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "string"
      tags:
        - audience
    delete:
      summary: "DELETE api/v1/admin/audience"
      operationId: "deleteAudience"
      parameters:
        - name: "name"
          in: "query"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "string"
      tags:
        - audience
  /api/v1/accounts/login:
    get:
      summary: "GET api/v1/accounts/login"
      operationId: "login"
      parameters:
        - name: "Authorization"
          in: "header"
          required: true
          schema:
            type: "string"
        - name: "role"
          in: "header"
          required: true
          schema:
            type: "string"
            enum:
              - "ROLE_ADMIN"
              - "ROLE_USER"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LoginDto"
      tags:
        - auth
  /api/v1/accounts/token/refresh:
    post:
      summary: "POST api/v1/accounts/token/refresh"
      operationId: "refreshToken"
      parameters:
        - name: "Authorization"
          in: "header"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LoginDto"
      tags:
        - auth
...
components:
  schemas:
    AudienceDTO:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        name:
          type: "string"
        audienceGroupCapacity:
          type: "integer"
          format: "int64"
...
    RoleDto:
      type: "object"
      properties:
        name:
          type: "string"
          enum:
            - "ROLE_ADMIN"
            - "ROLE_USER"
    LoginDto:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        username:
          type: "string"
        login:
          type: "string"
        accessToken:
          type: "string"
        refreshToken:
          type: "string"
        roles:
          type: "array"
          items:
            $ref: "#/components/schemas/RoleDto"
...